+++ frontend/soltravel/src/redux/store.tsx
@@ -1,0 +1,12 @@
+import { configureStore } from "@reduxjs/toolkit";
+import accountReducer from "./accountSlice";
+
+export const store = configureStore({
+  reducer: {
+    account: accountReducer,
+  },
+});
+
+export type RootState = ReturnType<typeof store.getState>;
+
+export type AppDispatch = typeof store.dispatch;
